+++ src/main/java/org/basicData/model/CarGroup.java
@@ -26,5 +26,5 @@ public class CarGroup extends BaseEntity {
     @Column(columnDefinition = "decimal(18, 0)", name = "f_company_id")
     private Long companyId;
     @Column(columnDefinition = "float", name = "factor_value")
-    private Float factorValue;
+    private Double factorValue;
 }
